Fixed Prometheus. Now he can follow the correct rules of the game.
To do that a method in the GodController has been edited (askBuildAgain) and a methiod in the GodView have been added to correctly interact with the owner.
New methods in the WorkerMap and in the WorkerMoveMap have been added to distinguish the situation where the worker can a priori reach an higher level or not. This would change the evolution of the turn on the decision the player can make.