Added a way to handle client messages when it's not their turn.
So ViewClient implements Runnable and handles out of turn requests.
NetworkHandler has now a boolean flag (maybe to edit) to highlight if it's his client turn or not.
In the CLIView there are now methods to print the startin point and the end of a turn, moreover the notYourTurn method.
The methods to start the turn and to end it are handled by the TurnHandler.